tree-wide: introduce macros to define *_ref() and *_unref() functions

+++ src/libsystemd/sd-device/device-internal.h
@@ -7,7 +7,7 @@
 #include "set.h"
 
 struct sd_device {
-        uint64_t n_ref;
+        unsigned n_ref;
 
         sd_device *parent;
         bool parent_set; /* no need to try to reload parent */
